Bytecode: Initial code breaking modularization
This WILL break your code. All of it.
Anyways, this attempts to put Lua 5.1, Lua 5.2 and Lua 5.3 into their
own seperate modules. When running on 5.1, you can only work with Lua
5.1 bytecode; on 5.2 you can work with 5.1 and 5.2 bytecode, and on 5.3
you can work with 5.1, 5.2, and 5.3 bytecode.
